api/requirements.txt
panda
sidrapy